+++ sdk/identity/azure_identity/src/lib.rs
@@ -19,6 +19,7 @@ mod env;
 mod imds_managed_identity_credential;
 mod managed_identity_credential;
 mod process;
+mod secret_bytes;
 mod virtual_machine_managed_identity_credential;
 mod workload_identity_credential;
 
@@ -32,6 +33,7 @@ pub use client_secret_credential::*;
 pub use developer_tools_credential::*;
 pub use managed_identity_credential::*;
 pub use process::{new_executor, Executor};
+pub use secret_bytes::SecretBytes;
 pub use workload_identity_credential::*;
 
 pub(crate) use app_service_managed_identity_credential::*;